1.请求参数为固定参数 2.请求设置本地缓存页面开关 3.返回参数默认值

+++ document_sdk/iospan/static/long.h
@@ -1,2 +1,2 @@
-<br />/** {note} */
-<br />static const long {!} = {?};
+<br />/** {note} 长整形Long*/
+<br />static const NSInteger {!} = {?};
